Add files via upload
Initial draft for Reports and Inventory pages. 

Since the draft was made, we received some feedback from the Client about our Project Plan. Changes have been made on the Figma since so it may not match the design draft exactly. Additionally, after another visit to the client's site, we decided to implement an additional Checkout page. This page was made purely on Figma.